RavenDB-14706 move Initialize & varchar stuff to SqlDocumentTransformer from RelationalDatabaseDocumentTransformerBase
